Add support to iframe chatbot
