+++ api/apps/conversation_app.py
@@ -17,12 +17,14 @@ import json
 import re
 import traceback
 from copy import deepcopy
+
+from api.db.services.conversation_service import ConversationService
 from api.db.services.user_service import UserTenantService
 from flask import request, Response
 from flask_login import login_required, current_user
 
 from api.db import LLMType
-from api.db.services.dialog_service import DialogService, ConversationService, chat, ask
+from api.db.services.dialog_service import DialogService, chat, ask
 from api.db.services.knowledgebase_service import KnowledgebaseService
 from api.db.services.llm_service import LLMBundle, TenantService, TenantLLMService
 from api import settings
