+++ api/db/services/canvas_service.py
@@ -13,10 +13,15 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 #
-from datetime import datetime
-import peewee
-from api.db.db_models import DB, API4Conversation, APIToken, Dialog, CanvasTemplate, UserCanvas
+import json
+from copy import deepcopy
+from uuid import uuid4
+from agent.canvas import Canvas
+from api.db.db_models import DB, API4Conversation, CanvasTemplate, UserCanvas
+from api.db.services.api_service import API4ConversationService
 from api.db.services.common_service import CommonService
+from api.db.services.conversation_service import structure_answer
+from api.utils import get_uuid
 
 
 class CanvasTemplateService(CommonService):
@@ -44,3 +49,113 @@ class UserCanvasService(CommonService):
         agents = agents.paginate(page_number, items_per_page)
 
         return list(agents.dicts())
+
+
+def completion(tenant_id, agent_id, question, session_id=None, stream=True, **kwargs):
+    e, cvs = UserCanvasService.get_by_id(agent_id)
+    assert e, "Agent not found."
+    assert cvs.user_id == tenant_id, "You do not own the agent."
+
+    if not isinstance(cvs.dsl, str):
+        cvs.dsl = json.dumps(cvs.dsl, ensure_ascii=False)
+    canvas = Canvas(cvs.dsl, tenant_id)
+
+    if not session_id:
+        session_id = get_uuid()
+        conv = {
+            "id": session_id,
+            "dialog_id": cvs.id,
+            "user_id": kwargs.get("user_id", ""),
+            "message": [{"role": "assistant", "content": canvas.get_prologue()}],
+            "source": "agent",
+            "dsl": json.loads(cvs.dsl)
+        }
+        API4ConversationService.save(**conv)
+        yield "data:" + json.dumps({"code": 0,
+                                    "message": "",
+                                    "data": {
+                                        "session_id": session_id,
+                                        "answer": canvas.get_prologue(),
+                                        "reference": [],
+                                        "param": canvas.get_preset_param()
+                                      }
+                                    },
+                                   ensure_ascii=False) + "\n\n"
+        yield "data:" + json.dumps({"code": 0, "message": "", "data": True}, ensure_ascii=False) + "\n\n"
+        return
+    else:
+        session_id = session_id
+        e, conv = API4ConversationService.get_by_id(session_id)
+        assert e, "Session not found!"
+        canvas = Canvas(json.dumps(conv.dsl), tenant_id)
+
+    messages = conv.message
+    question = {
+        "role": "user",
+        "content": question,
+        "id": str(uuid4())
+    }
+    messages.append(question)
+    msg = []
+    for m in messages:
+        if m["role"] == "system":
+            continue
+        if m["role"] == "assistant" and not msg:
+            continue
+        msg.append(m)
+    if not msg[-1].get("id"): msg[-1]["id"] = get_uuid()
+    message_id = msg[-1]["id"]
+
+    if not conv.reference:
+        conv.reference = []
+    conv.message.append({"role": "assistant", "content": "", "id": message_id})
+    conv.reference.append({"chunks": [], "doc_aggs": []})
+
+    final_ans = {"reference": [], "content": ""}
+
+    canvas.add_user_input(msg[-1]["content"])
+
+    if stream:
+        try:
+            for ans in canvas.run(stream=stream):
+                if ans.get("running_status"):
+                    yield "data:" + json.dumps({"code": 0, "message": "",
+                                                "data": {"answer": ans["content"],
+                                                         "running_status": True}},
+                                               ensure_ascii=False) + "\n\n"
+                    continue
+                for k in ans.keys():
+                    final_ans[k] = ans[k]
+                ans = {"answer": ans["content"], "reference": ans.get("reference", [])}
+                ans = structure_answer(conv, ans, message_id, session_id)
+                yield "data:" + json.dumps({"code": 0, "message": "", "data": ans},
+                                           ensure_ascii=False) + "\n\n"
+
+            canvas.messages.append({"role": "assistant", "content": final_ans["content"], "id": message_id})
+            canvas.history.append(("assistant", final_ans["content"]))
+            if final_ans.get("reference"):
+                canvas.reference.append(final_ans["reference"])
+            conv.dsl = json.loads(str(canvas))
+            API4ConversationService.append_message(conv.id, conv.to_dict())
+        except Exception as e:
+            conv.dsl = json.loads(str(canvas))
+            API4ConversationService.append_message(conv.id, conv.to_dict())
+            yield "data:" + json.dumps({"code": 500, "message": str(e),
+                                        "data": {"answer": "**ERROR**: " + str(e), "reference": []}},
+                                       ensure_ascii=False) + "\n\n"
+        yield "data:" + json.dumps({"code": 0, "message": "", "data": True}, ensure_ascii=False) + "\n\n"
+
+    else:
+        for answer in canvas.run(stream=False):
+            if answer.get("running_status"): continue
+            final_ans["content"] = "\n".join(answer["content"]) if "content" in answer else ""
+            canvas.messages.append({"role": "assistant", "content": final_ans["content"], "id": message_id})
+            if final_ans.get("reference"):
+                canvas.reference.append(final_ans["reference"])
+            conv.dsl = json.loads(str(canvas))
+
+            result = {"answer": final_ans["content"], "reference": final_ans.get("reference", [])}
+            result = structure_answer(conv, result, message_id, session_id)
+            API4ConversationService.append_message(conv.id, conv.to_dict())
+            yield result
+            break
