+++ api/db/services/dialog_service.py
@@ -23,7 +23,7 @@ from timeit import default_timer as timer
 import datetime
 from datetime import timedelta
 from api.db import LLMType, ParserType,StatusEnum
-from api.db.db_models import Dialog, Conversation,DB
+from api.db.db_models import Dialog, DB
 from api.db.services.common_service import CommonService
 from api.db.services.knowledgebase_service import KnowledgebaseService
 from api.db.services.llm_service import LLMService, TenantLLMService, LLMBundle
@@ -58,27 +58,6 @@ class DialogService(CommonService):
         chats = chats.paginate(page_number, items_per_page)
 
         return list(chats.dicts())
-
-
-class ConversationService(CommonService):
-    model = Conversation
-
-    @classmethod
-    @DB.connection_context()
-    def get_list(cls,dialog_id,page_number, items_per_page, orderby, desc, id , name):
-        sessions = cls.model.select().where(cls.model.dialog_id ==dialog_id)
-        if id:
-            sessions = sessions.where(cls.model.id == id)
-        if name:
-            sessions = sessions.where(cls.model.name == name)
-        if desc:
-            sessions = sessions.order_by(cls.model.getter_by(orderby).desc())
-        else:
-            sessions = sessions.order_by(cls.model.getter_by(orderby).asc())
-
-        sessions = sessions.paginate(page_number, items_per_page)
-
-        return list(sessions.dicts())
 
 
 def message_fit_in(msg, max_length=4000):
